move GenericResponse to reponse package
